feature/task-refactor: feat: :sparkles: define the logic for the flow to create the release from the pullrequest
