+++ tasks/shared/utils/index.ts
@@ -3,7 +3,7 @@ export * from './getPullRequestId'
 export * from './getReleaseType'
 export * from './getReleaseVersion'
 export * from './git'
-export * from './initialSetup'
+export * from './setup'
 export * from './manifestUtils'
 export * from './markdown'
 export * from './messages'
