+++ tasks/shared/utils/messages/vtex/index.ts
@@ -1,0 +1,1 @@
+export * from './publish';
