+++ tasks/shared/utils/messages/vtex/publish/index.ts
@@ -1,0 +1,2 @@
+export * from './beta'
+export * from './publish'
